yt_shorts_automation/requirements.txt
google-genai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
openai-whisper
moviepy
python-dotenv
openai
requests
